scripts/requirements.txt
# For scripts/prepare-favicon.py (favicon and logo generation)
Pillow>=10.0.0
